feat(KAMP-480): new-arrival highlight for list-view album rows
Extract highlight state machine into shared useNewArrivalHighlight hook,
consumed by both AlbumCard and ListRow. Add all 7 highlight styles to list
rows with geometry adapted for the 48px-tall thumbnail format.
