+++ kamp_ui/src/renderer/src/assets/modules.css
@@ -96,6 +96,7 @@
 }
 
 .module-list-row {
+  position: relative;
   display: flex;
   align-items: center;
   gap: 10px;
