+++ kamp_ui/src/renderer/src/assets/new-arrival-highlight.css
@@ -520,3 +520,250 @@
     display: none;
   }
 }
+
+/* ── List-row highlight variants ────────────────────────────────────────── */
+/* Modifier class: module-list-row--highlight-<name>                        */
+/* Geometry adapts for the 48px-tall list row vs the ~255px album card.     */
+
+/* ── shiny (list row) ───────────────────────────────────────────────────── */
+
+.module-list-row--highlight-shiny {
+  animation: shiny-aura 3.2s ease-in-out infinite;
+}
+
+.module-list-row--highlight-shiny:hover {
+  box-shadow:
+    0 4px 12px rgba(0, 0, 0, 0.4),
+    0 0 8px 2px rgba(255, 215, 0, 0.22);
+  animation: none;
+}
+
+.module-list-row.is-mounting .shiny-sweep {
+  animation: shiny-sweep-once 1.2s cubic-bezier(0.4, 0, 0.2, 1) 1;
+}
+
+.module-list-row--highlight-shiny:hover:not(.is-mounting) .shiny-sweep {
+  animation: shiny-sweep-fast 1.8s cubic-bezier(0.4, 0, 0.2, 1) infinite;
+}
+
+/* ── newmoji (list row) ─────────────────────────────────────────────────── */
+
+/* module-list-info needs position:relative so the badge can use absolute positioning */
+.module-list-row--highlight-newmoji .module-list-info {
+  position: relative;
+}
+
+/* Badge smaller (14px vs 18px) to fit the 48px row height */
+.module-list-row--highlight-newmoji .newmoji-badge {
+  font-size: 14px;
+  top: -8px;
+  left: 4px;
+}
+
+.module-list-row--highlight-newmoji:hover .newmoji-badge {
+  animation:
+    newmoji-idle 5s ease-in-out infinite,
+    newmoji-hover-scale 0.45s cubic-bezier(0.34, 1.56, 0.64, 1) 1 forwards;
+}
+
+/* ── boring (list row) ──────────────────────────────────────────────────── */
+
+/* Ribbon scaled down for 48px thumb: width 80px, top 14px, right -22px.
+   The geometry comment on album-card: W=180, top=23, right=-28, width=120.
+   Scale factor ≈ 48/180 × 120 ≈ 32 (width), adjusted empirically for clean clip. */
+.module-list-row--highlight-boring .module-list-thumb::after,
+.module-list-row--highlight-boring .boring-hover {
+  position: absolute;
+  top: 14px;
+  right: -22px;
+  width: 80px;
+  padding: 4px 0;
+  line-height: 1;
+  background: #3B6FD4;
+  color: #fff;
+  font-family: inherit;
+  font-size: 7px;
+  font-weight: 700;
+  letter-spacing: 0.12em;
+  text-align: center;
+  transform: rotate(45deg);
+  pointer-events: none;
+  z-index: 2;
+  transition: opacity 80ms ease;
+}
+
+.module-list-row--highlight-boring .module-list-thumb::after {
+  content: 'New';
+}
+
+.module-list-row--highlight-boring .boring-hover {
+  opacity: 0;
+}
+
+.module-list-row--highlight-boring:hover .module-list-thumb::after {
+  opacity: 0;
+}
+
+.module-list-row--highlight-boring:hover .boring-hover {
+  opacity: 1;
+}
+
+/* ── vaporwave (list row) ───────────────────────────────────────────────── */
+
+.module-list-row--highlight-vaporwave {
+  animation: vaporwave-halo 6s linear infinite;
+}
+
+.module-list-row--highlight-vaporwave:hover {
+  animation: vaporwave-halo 6s linear infinite;
+  filter: brightness(1.08);
+}
+
+.module-list-row--highlight-vaporwave:hover .vaporwave-scanlines {
+  opacity: 0;
+}
+
+/* ── proud (list row) ───────────────────────────────────────────────────── */
+
+/* Same dual background-image + border-box technique as album-card--highlight-proud.
+   .module-list-row:hover uses `background` shorthand — redeclare background-image below. */
+.module-list-row--highlight-proud {
+  border: 3px solid transparent;
+  background-color: var(--surface);
+  background-image:
+    linear-gradient(var(--surface), var(--surface)),
+    repeating-linear-gradient(
+      135deg,
+      #FFFFFF   0px,  #FFFFFF  12px,
+      #F7A8B8  12px,  #F7A8B8  24px,
+      #55CDFC  24px,  #55CDFC  36px,
+      #C47A2B  36px,  #C47A2B  48px,
+      #E40303  48px,  #E40303  60px,
+      #FF8C00  60px,  #FF8C00  72px,
+      #FFED00  72px,  #FFED00  84px,
+      #008026  84px,  #008026  96px,
+      #004DFF  96px,  #004DFF 108px,
+      #750787 108px,  #750787 120px
+    );
+  background-origin: padding-box, border-box;
+  background-clip: padding-box, border-box;
+  background-size: auto, 170px 170px;
+  box-shadow: inset 0 0 8px 1px rgba(85, 205, 252, 0.15);
+  animation: proud-barber 6s linear infinite;
+}
+
+.module-list-row--highlight-proud:hover {
+  background-color: var(--surface-hover);
+  background-image:
+    linear-gradient(var(--surface-hover), var(--surface-hover)),
+    repeating-linear-gradient(
+      135deg,
+      #FFFFFF   0px,  #FFFFFF  12px,
+      #F7A8B8  12px,  #F7A8B8  24px,
+      #55CDFC  24px,  #55CDFC  36px,
+      #C47A2B  36px,  #C47A2B  48px,
+      #E40303  48px,  #E40303  60px,
+      #FF8C00  60px,  #FF8C00  72px,
+      #FFED00  72px,  #FFED00  84px,
+      #008026  84px,  #008026  96px,
+      #004DFF  96px,  #004DFF 108px,
+      #750787 108px,  #750787 120px
+    );
+  background-origin: padding-box, border-box;
+  background-clip: padding-box, border-box;
+  background-size: auto, 170px 170px;
+  animation-duration: 1.6s;
+}
+
+/* ── pressed (list row) ─────────────────────────────────────────────────── */
+
+.module-list-row--highlight-pressed {
+  animation: pressed-breath 4s ease-in-out infinite;
+}
+
+.module-list-row--highlight-pressed:hover .pressed-glint-hover {
+  animation: pressed-glint-once 0.5s ease-in-out forwards;
+}
+
+/* ── static (list row) ──────────────────────────────────────────────────── */
+
+/* Same dual background-image + --static-border-gradient technique.
+   .module-list-row:hover shorthand resets background-image — redeclare below. */
+.module-list-row--highlight-static {
+  border: 3px solid transparent;
+  background-color: var(--surface);
+  background-image:
+    linear-gradient(var(--surface), var(--surface)),
+    var(--static-border-gradient, conic-gradient(from 180deg, #bbb 0deg, #999 180deg, #bbb 360deg));
+  background-origin: padding-box, border-box;
+  background-clip: padding-box, border-box;
+}
+
+.module-list-row--highlight-static:hover {
+  background-color: var(--surface-hover);
+  background-image:
+    linear-gradient(var(--surface-hover), var(--surface-hover)),
+    var(--static-border-gradient, conic-gradient(from 180deg, #bbb 0deg, #999 180deg, #bbb 360deg));
+  background-origin: padding-box, border-box;
+  background-clip: padding-box, border-box;
+}
+
+.module-list-row--highlight-static:hover .static-sparks,
+.module-list-row--highlight-static:hover .static-aura {
+  opacity: 0;
+}
+
+/* ── Reduced motion: list-row variants ──────────────────────────────────── */
+
+@media (prefers-reduced-motion: reduce) {
+  .module-list-row--highlight-shiny {
+    animation: none;
+    box-shadow: none;
+    outline: 2px solid rgba(255, 215, 0, 0.6);
+    outline-offset: 0;
+  }
+
+  .module-list-row--highlight-newmoji .newmoji-badge {
+    animation: none;
+    translate: 0 0;
+    rotate: 0deg;
+    scale: 1;
+  }
+
+  .module-list-row--highlight-vaporwave {
+    animation: none;
+    box-shadow: 0 0 0 2px #A855F7;
+  }
+
+  .module-list-row--highlight-vaporwave .vaporwave-scanlines {
+    display: none;
+  }
+
+  .module-list-row--highlight-proud {
+    animation: none;
+  }
+
+  .module-list-row--highlight-pressed {
+    animation: none;
+    box-shadow: 0 0 0 1px rgba(255, 255, 255, 0.50);
+  }
+
+  .module-list-row--highlight-pressed .pressed-glint,
+  .module-list-row--highlight-pressed .pressed-glint-hover {
+    display: none;
+  }
+
+  .module-list-row--highlight-static {
+    --static-border-gradient: conic-gradient(
+      from 180deg,
+      #bbb 0deg,
+      #999 180deg,
+      #bbb 360deg
+    ) !important;
+  }
+
+  .module-list-row--highlight-static .static-spark,
+  .module-list-row--highlight-static .static-aura {
+    display: none;
+  }
+}
